+++ packages/contracts/contracts/voting/multisig/MultisigSetup.sol
@@ -25,7 +25,7 @@ contract MultisigSetup is PluginSetup {
 
     /// @inheritdoc IPluginSetup
     function prepareInstallationDataABI() external pure returns (string memory) {
-        return "(uint256 minApprovals, address[] members)";
+        return "(address[] members, tuple(bool onlyListed, uint16 minApprovals))";
     }
 
     /// @inheritdoc IPluginSetup
@@ -40,12 +40,20 @@ contract MultisigSetup is PluginSetup {
         IDAO dao = IDAO(_dao);
 
         // Decode `_data` to extract the params needed for deploying and initializing `Multisig` plugin.
-        (uint256 minApprovals, address[] memory members) = abi.decode(_data, (uint256, address[]));
+        (
+            address[] memory members,
+            Multisig.MultisigSettings memory multisigSettings
+        ) = abi.decode(_data, (address[], Multisig.MultisigSettings));
 
         // Prepare and Deploy the plugin proxy.
         plugin = createERC1967Proxy(
             address(multisigBase),
-            abi.encodeWithSelector(Multisig.initialize.selector, dao, minApprovals, members)
+            abi.encodeWithSelector(
+                Multisig.initialize.selector,
+                dao,
+                members,
+                multisigSettings
+            )
         );
 
         // Prepare helpers