+++ packages/contracts/test/voting/multisig.ts
@@ -3,7 +3,12 @@ import {ethers} from 'hardhat';
 import {SignerWithAddress} from '@nomiclabs/hardhat-ethers/signers';
 
 import {DAO} from '../../typechain';
-import {findEvent, DAO_EVENTS, VOTING_EVENTS} from '../../utils/event';
+import {
+  findEvent,
+  DAO_EVENTS,
+  VOTING_EVENTS,
+  MULTISIG_EVENTS,
+} from '../../utils/event';
 import {getMergedABI} from '../../utils/abi';
 import {customError, ERRORS} from '../test-utils/custom-error-helper';
 
@@ -15,6 +20,7 @@ describe('Multisig', function () {
   let dummyMetadata: string;
 
   let minApprovals: number;
+  let onlyListed: boolean;
 
   const id = 0;
 
@@ -53,6 +59,7 @@ describe('Multisig', function () {
 
   beforeEach(async () => {
     minApprovals = 3;
+    onlyListed = true;
 
     const MultisigFactory = new ethers.ContractFactory(
       mergedAbi,
@@ -85,10 +92,16 @@ describe('Multisig', function () {
 
   describe('initialize: ', async () => {
     it('reverts if trying to re-initialize', async () => {
-      await multisig.initialize(dao.address, minApprovals, addresslist(5));
+      await multisig.initialize(dao.address, addresslist(5), {
+        onlyListed,
+        minApprovals,
+      });
 
       await expect(
-        multisig.initialize(dao.address, minApprovals, addresslist(5))
+        multisig.initialize(dao.address, addresslist(5), {
+          onlyListed,
+          minApprovals,
+        })
       ).to.be.revertedWith(ERRORS.ALREADY_INITIALIZED);
     });
   });
@@ -96,7 +109,10 @@ describe('Multisig', function () {
   describe('Addresslisting members: ', async () => {
     it('should return false, if a user is not listed', async () => {
       minApprovals = 1;
-      await multisig.initialize(dao.address, minApprovals, addresslist(1));
+      await multisig.initialize(dao.address, addresslist(1), {
+        onlyListed,
+        minApprovals,
+      });
 
       const block1 = await ethers.provider.getBlock('latest');
       await ethers.provider.send('evm_mine', []);
@@ -107,7 +123,10 @@ describe('Multisig', function () {
 
     it('should add new members to the address list', async () => {
       minApprovals = 1;
-      await multisig.initialize(dao.address, minApprovals, addresslist(1));
+      await multisig.initialize(dao.address, addresslist(1), {
+        onlyListed,
+        minApprovals,
+      });
 
       const block1 = await ethers.provider.getBlock('latest');
       await ethers.provider.send('evm_mine', []);
@@ -133,7 +152,10 @@ describe('Multisig', function () {
 
     it('should remove users from the address list', async () => {
       minApprovals = 1;
-      await multisig.initialize(dao.address, minApprovals, addresslist(2));
+      await multisig.initialize(dao.address, addresslist(2), {
+        onlyListed,
+        minApprovals,
+      });
 
       const block1 = await ethers.provider.getBlock('latest');
       await ethers.provider.send('evm_mine', []);
@@ -159,7 +181,10 @@ describe('Multisig', function () {
 
     it('reverts if the address list would become empty', async () => {
       minApprovals = 1;
-      await multisig.initialize(dao.address, minApprovals, addresslist(1));
+      await multisig.initialize(dao.address, addresslist(1), {
+        onlyListed,
+        minApprovals,
+      });
 
       await expect(
         multisig.removeAddresses([signers[0].address])
@@ -174,7 +199,10 @@ describe('Multisig', function () {
 
     it('reverts if the address list would become shorter than the current minimum approval parameter requires', async () => {
       minApprovals = 2;
-      await multisig.initialize(dao.address, minApprovals, addresslist(3));
+      await multisig.initialize(dao.address, addresslist(3), {
+        onlyListed,
+        minApprovals,
+      });
 
       await expect(multisig.removeAddresses([signers[1].address])).to.not.be
         .reverted;
@@ -196,11 +224,11 @@ describe('Multisig', function () {
       minApprovals = 1;
     });
 
-    it('reverts if the user is not allowed to create a proposal', async () => {
+    it('reverts if the user is not on the list and only listed accounts can create proposals', async () => {
       await multisig.initialize(
         dao.address,
-        minApprovals,
-        addresslist(1) // signers[0] is listed
+        addresslist(1), // signers[0] is listed
+        {onlyListed: true, minApprovals}
       );
 
       await expect(
@@ -218,8 +246,27 @@ describe('Multisig', function () {
       ).to.not.be.reverted;
     });
 
+    it('creates a proposal when unlisted accounts are allowed', async () => {
+      await multisig.initialize(
+        dao.address,
+        addresslist(1), // signer[0] is listed
+        {onlyListed: false, minApprovals}
+      );
+
+      expect(
+        await multisig
+          .connect(signers[1])
+          .createProposal(dummyMetadata, dummyActions, false, false)
+      )
+        .to.emit(multisig, VOTING_EVENTS.PROPOSAL_CREATED)
+        .withArgs(id, signers[0].address, dummyMetadata);
+    });
+
     it('creates a proposal successfully and does not approve if not specified', async () => {
-      await multisig.initialize(dao.address, minApprovals, addresslist(1));
+      await multisig.initialize(dao.address, addresslist(1), {
+        onlyListed,
+        minApprovals,
+      });
 
       expect(
         await multisig.createProposal(dummyMetadata, dummyActions, false, false)
@@ -248,7 +295,10 @@ describe('Multisig', function () {
     });
 
     it('creates a proposal successfully and approves if specified', async () => {
-      await multisig.initialize(dao.address, minApprovals, addresslist(1));
+      await multisig.initialize(dao.address, addresslist(1), {
+        onlyListed,
+        minApprovals,
+      });
 
       expect(
         await multisig.createProposal(dummyMetadata, dummyActions, true, false)
@@ -271,7 +321,10 @@ describe('Multisig', function () {
 
   describe('Proposal + Execute:', async () => {
     beforeEach(async () => {
-      await multisig.initialize(dao.address, minApprovals, addresslist(10));
+      await multisig.initialize(dao.address, addresslist(10), {
+        onlyListed,
+        minApprovals,
+      });
 
       expect(
         (
@@ -356,6 +409,69 @@ describe('Multisig', function () {
       await expect(multisig.execute(id)).to.be.revertedWith(
         customError('ProposalExecutionForbidden', id)
       );
+    });
+  });
+
+  describe('MultisigSettings', async () => {
+    it('should set the right minApprovals during initialization', async () => {
+      await multisig.initialize(dao.address, addresslist(5), {
+        onlyListed,
+        minApprovals,
+      });
+      expect((await multisig.multisigSettings()).minApprovals).to.be.eq(
+        minApprovals
+      );
+    });
+
+    it('should set the right onlyListed during initialization', async () => {
+      await multisig.initialize(dao.address, addresslist(5), {
+        onlyListed,
+        minApprovals,
+      });
+      expect((await multisig.multisigSettings()).onlyListed).to.be.eq(
+        onlyListed
+      );
+    });
+
+    it('should emit MultisigSettingsUpdated during initialization', async () => {
+      await expect(
+        multisig.initialize(dao.address, addresslist(5), {
+          onlyListed,
+          minApprovals,
+        })
+      )
+        .to.emit(multisig, MULTISIG_EVENTS.MULTISIG_SETTINGS_UPDATED)
+        .withArgs(onlyListed, minApprovals);
+    });
+
+    it('should not allow to set minApprovals higher than addresslist length', async () => {
+      await multisig.initialize(dao.address, addresslist(5), {
+        onlyListed,
+        minApprovals,
+      });
+      await expect(
+        multisig.updateMultisigSettings({onlyListed, minApprovals: 6})
+      ).to.be.revertedWith(customError('MinApprovalsOutOfBounds', 5, 6));
+    });
+
+    it('should not allow to set minApprovals would be set to 0', async () => {
+      await multisig.initialize(dao.address, addresslist(5), {
+        onlyListed,
+        minApprovals,
+      });
+      await expect(
+        multisig.updateMultisigSettings({onlyListed, minApprovals: 0})
+      ).to.be.revertedWith(customError('MinApprovalsOutOfBounds', 1, 0));
+    });
+
+    it('should emit MultisigSettingsUpdated when updateMutlsigSettings gets called', async () => {
+      await multisig.initialize(dao.address, addresslist(5), {
+        onlyListed,
+        minApprovals,
+      });
+      await expect(multisig.updateMultisigSettings({onlyListed, minApprovals}))
+        .to.emit(multisig, MULTISIG_EVENTS.MULTISIG_SETTINGS_UPDATED)
+        .withArgs(onlyListed, minApprovals);
     });
   });
 });