+++ packages/subgraph/CHANGELOG.md
@@ -17,6 +17,7 @@ On 2022-12-09 15:16:22
 - Fixed the early execution criterion in `MajorityVoting` by calculating the `worstCaseSupport` quantity.
 - Adds support for `PluginRepo`, `PluginRegistry` and `PluginSetupProcessor`
 - Added `Withdrawn`, `TrustedForwarderSet`, `StandardCallbackRegistered` and `handleStandardCallbackRegistered` events handling to `DaoTemplate`
+- Added support for the new `PluginSettingsUpdated` event of the `Multisig` plugin
 
 ### Changed
 
@@ -27,6 +28,7 @@ On 2022-12-09 15:16:22
 - Improved test recompilation.
 - Marks some entity as immutable.
 - Fixes calcuation crash in erc20 voting, when no votes were cast
+- Added the field `onlyListed` to the `MultisigPlugin` type
 
 ## 0.4.0-alpha